+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 4. add headers option
 5. set default userAgent as "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:128.0) Gecko/20100101 Firefox/128.0"
 6. fix some bugs
+7. remove `segmentDownloaded` event
 
 # 0.1.0